Added appropriate comments to the classes

+++ src/main/java/com/project/FileConvertor/Model/FileStorageProperties.java
@@ -4,12 +4,19 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 
 import lombok.Data;
 
+/**
+ * Represents file properties from yaml.
+ * This holds file storage info like base directory, maximum file size, etc.
+ */
 @Data
 @ConfigurationProperties("app.file.convertor")
 public class FileStorageProperties {
 
     // We can add other properties as needed in the porject
 
+    // Base directory of the folder where file is stored.
     private String baseDir;
-    private Integer maxSize;
+
+    // Max Size of a file that can be stored.
+    private Long maxSize;
 }
